+++ NEWS.md
@@ -1,3 +1,23 @@
+# ResultModelManager 0.5.0
+Changes:
+1. Added utility function `grantTablePermissions` to make it easier to grant select, delete, insert and update
+permissions for users on results database setups
+
+2. Added `ResultExportManager` class and utility to support standardized validation routines for exported data
+
+3. Allow packages to have an internal migration table prefix separate from the user defined one e.g `my_study_sccs_migration`
+
+4. `emptyIsNa` field is no longer required in specifications
+
+5. Purge all data supported in upload functionality (requiring user input)
+
+Bug fixes:
+
+1.  `generateSqlSchema` No longer requires primary key field to be lower case in results spec files, is now case-insensitive
+
+2. connectionHandlers now check sql string attribute to see if query needs translating, avoiding potential for errors
+caused by double translation
+
 # ResultModelManager 0.4.0
 Changes: